Make BlockState int

+++ Frontend/BlockState.cs
@@ -2,6 +2,6 @@ namespace Frontend
 {
     public struct BlockState
     {
-        public byte State;
+        public int State;
     }
 }
